Add conditional that I forgot in last commit

+++ JDMallen.Toolbox.Infrastructure.EFCore/Extensions/ServiceExtensions.cs
@@ -76,46 +76,49 @@ namespace Microsoft.Extensions.DependencyInjection
 		public static void DropTablesAndEnsureCreated(
 			this DbContext dbContext, 
 			bool dropTables = true)
-		{			
-			var conn = dbContext.Database.GetDbConnection();
-			conn.Open();
-			dbContext.Model.GetEntityTypes()
-				.ToList()
-				.ForEach(
-					et =>
-					{
-						bool tableExists;
-						var tableName =
-							et.GetAnnotations()
-							.FirstOrDefault(x => x.Name == "Relational:TableName")
-								?.Value.ToString()
-								.ToLowerInvariant()
-							?? et.ClrType.Name.ToLowerInvariant();
-						using (var cmd = conn.CreateCommand())
+		{
+			if (dropTables)
+			{
+				var conn = dbContext.Database.GetDbConnection();
+				conn.Open();
+				dbContext.Model.GetEntityTypes()
+					.ToList()
+					.ForEach(
+						et =>
 						{
-							cmd.CommandText =
-								"SELECT COUNT(*) FROM information_schema.TABLES " +
-								"WHERE TABLE_SCHEMA = " + 
-								$"\'{dbContext.Database.GetDbConnection().Database}\' " +
-								$"AND TABLE_NAME = \'{tableName.ToLowerInvariant()}\';";
-							cmd.CommandType = CommandType.Text;
-							using (var reader = cmd.ExecuteReader())
+							bool tableExists;
+							var tableName =
+								et.GetAnnotations()
+								.FirstOrDefault(x => x.Name == "Relational:TableName")
+									?.Value.ToString()
+									.ToLowerInvariant()
+								?? et.ClrType.Name.ToLowerInvariant();
+							using (var cmd = conn.CreateCommand())
 							{
-								reader.Read();
-								tableExists = reader.GetInt32(0) > 0;
-								reader.Close();
+								cmd.CommandText =
+									"SELECT COUNT(*) FROM information_schema.TABLES " +
+									"WHERE TABLE_SCHEMA = " + 
+									$"\'{dbContext.Database.GetDbConnection().Database}\' " +
+									$"AND TABLE_NAME = \'{tableName.ToLowerInvariant()}\';";
+								cmd.CommandType = CommandType.Text;
+								using (var reader = cmd.ExecuteReader())
+								{
+									reader.Read();
+									tableExists = reader.GetInt32(0) > 0;
+									reader.Close();
+								}
 							}
-						}
-						if (!tableExists)
-							return;
-						var dropCommand = 
-							"SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, " +
-							"FOREIGN_KEY_CHECKS=0;" +
-							$"DROP TABLE `{tableName.ToLowerInvariant()}`;" +
-							"SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;";
-						dbContext.Database.ExecuteSqlCommand(dropCommand);
-					});
-			conn.Close();
+							if (!tableExists)
+								return;
+							var dropCommand = 
+								"SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, " +
+								"FOREIGN_KEY_CHECKS=0;" +
+								$"DROP TABLE `{tableName.ToLowerInvariant()}`;" +
+								"SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;";
+							dbContext.Database.ExecuteSqlCommand(dropCommand);
+						});
+				conn.Close();
+			}
 			dbContext.Database.EnsureCreated();
 		}
